requirement.txt
langchain-openai
langchain_core
python-dotenv
streamlit
langchain-community
langserve
bs4
pypdf
faiss-cpu
